+++ nest/products/oda/endpoints.py
@@ -1,7 +1,7 @@
 from django.http import HttpRequest
 from ninja import Router, Schema
 
-from nest.api.fields import FormField
+from nest.forms.fields import FormField
 from nest.api.responses import APIResponse
 from nest.core.decorators import staff_required
