+++ nest/recipes/ingredients/endpoints.py
@@ -2,7 +2,7 @@ from django.http import HttpRequest
 from ninja import Router, Schema
 from store_kit.http import status
 
-from nest.api.fields import FormField
+from nest.forms.fields import FormField
 from nest.api.responses import APIResponse
 from nest.core.decorators import staff_required
 from nest.frontend.components import FrontendComponents
